Fix missing prop
